feat(pairwise): make bf16 the default compute ceiling
Default LUT payloads and GPU routing paths to BF16 while preserving FP32 margins and reduction accumulators. Extend Triton and TileLang routing to accept BF16/FP16 inputs, return gradients in the original input dtype, and align Zig and benchmark defaults with reduced-precision execution. Add explicit dtype coverage and retain FP32 only in precision-sensitive Walsh tests.

+++ backends/pairwise_zig.py
@@ -50,7 +50,7 @@ def pairwise_zig_forward(
     thresholds: Tensor,
     lut: Tensor,
     *,
-    lut_dtype: Literal["f32", "f16"] = "f32",
+    lut_dtype: Literal["f32", "f16"] = "f16",
 ) -> Tensor:
     if latent.device.type != "cpu":
         raise ValueError("PairwiseLUT backend='zig' requires CPU input tensors")
@@ -170,7 +170,7 @@ def pairwise_zig_soa_forward(
     thresholds: Tensor,
     lut: Tensor,
     *,
-    lut_dtype: Literal["f32", "f16"] = "f32",
+    lut_dtype: Literal["f32", "f16"] = "f16",
 ) -> Tensor:
     if latent.device.type != "cpu":
         raise ValueError("PairwiseLUT SoA Zig forward requires CPU input tensors")
@@ -244,7 +244,7 @@ def pairwise_zig_paged_forward(
     thresholds: Tensor,
     lut: Tensor,
     *,
-    lut_dtype: Literal["f32", "f16"] = "f32",
+    lut_dtype: Literal["f32", "f16"] = "f16",
     page_size: int = 1024,
 ) -> Tensor:
     if page_size < 1:
